+++ src/App.js
@@ -1,7 +1,10 @@
 import { lazy, Suspense } from "react";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import { createBrowserRouter, Navigate, RouterProvider } from "react-router-dom";
 import Root from "./Root.js";
 import RouteError from "./components/RouteError/RouteError.jsx";
+import Login from "./components/Login/Login.jsx";
+import Home from "./components/Main/Home/Home.jsx";
+import EmployeeContextProvider from "./store/employee-context.jsx";
 import "./App.css";
 
 const Metrics = lazy(() => import('./components/Main/Metrics/Metrics.jsx'));
@@ -15,13 +18,16 @@ const Reports = lazy(() => import('./components/Main/Reports/Reports.jsx'));
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <Root />,
+    element: <EmployeeContextProvider><Root /></EmployeeContextProvider>,
     errorElement: <RouteError root={true}/>,
     children: [
-      { index: true, element: <Suspense><Metrics /></Suspense>, errorElement: <RouteError />, loader: () => import('./components/Main/Metrics/Metrics.jsx').then(module => module.thresholdsLoader())},
+      { index: true, element: <Navigate to={'/home'} replace/>},
+      { path: '/home', element: <Home>starting page</Home>, errorElement: <RouteError />},
+      { path: '/metrics', element: <Suspense><Metrics /></Suspense>, errorElement: <RouteError />, loader: () => import('./components/Main/Metrics/Metrics.jsx').then(module => module.thresholdsLoader())},
       { path: '/selector', element: <Suspense><SelectorContextProvider><Selector /></SelectorContextProvider></Suspense>, errorElement: <RouteError />, loader: () => import('./components/Main/Selector/Selector.jsx').then(module => module.filtersLoader())},
       { path: '/company', element: <Suspense><CompanyContextProvider><Company /></CompanyContextProvider></Suspense>},
       { path: '/reports', element: <Suspense><ReportsContextProvider><Reports /></ReportsContextProvider></Suspense>, errorElement: <RouteError />, loader: () => import('./components/Main/Reports/Reports.jsx').then(module => module.reportsLoader())},
+      { path: '/metrics', element: <Login />, errorElement: <RouteError />},
     ],
   },
 ]);