+++ src/Root.js
@@ -1,14 +1,18 @@
-import { useState } from "react";
+import { useContext, useState } from "react";
 import { Outlet } from "react-router-dom";
 import Banner from "./components/Banner/Banner.jsx";
 import Main from "./components/Main/Main.jsx";
+import Footer from "./components/Footer/Footer.jsx";
+import Login from "./components/Login/Login.jsx";
 import { NavigationContext } from "./store/navigation-context.jsx";
+import { EmployeeContext } from "./store/employee-context.jsx";
 
 export default function Root() {
+  const { authenticated } = useContext(EmployeeContext);
   const [selectedNavItem, setSelectedNavItem] = useState(() => {
     const storedNavItem = JSON.parse(sessionStorage.getItem("selectedNavItem"));
     if (storedNavItem === undefined || storedNavItem === null) {
-      return "Metrics";
+      return "Home";
     } else {
       return storedNavItem;
     }
@@ -19,6 +23,14 @@ export default function Root() {
     setSelectedNavItem(() => selectedItem);
   }
 
+  function checkCurrentSession() {
+    const userData = JSON.parse(sessionStorage.getItem("userData"));
+    if (userData !== null && userData.authenticated === true) {
+      return true;
+    }
+    return false;
+  }
+
   const ctxNavigation = {
     selectedNavItem: selectedNavItem,
     selectNavItemHandler: selectNavItemHandler,
@@ -26,12 +38,17 @@ export default function Root() {
 
   return (
     <>
-      <NavigationContext.Provider value={ctxNavigation}>
-        <Banner />
-        <Main>
-          <Outlet />
-        </Main>
-      </NavigationContext.Provider>
+      {(authenticated === false && checkCurrentSession() === false) ?
+        <Login />
+        :
+        <NavigationContext.Provider value={ctxNavigation}>
+          <Banner />
+          <Main>
+            <Outlet />
+          </Main>
+          <Footer />
+        </NavigationContext.Provider>
+      }
     </>
   );
 }