+++ src/components/Main/Main.styles.jsx
@@ -3,6 +3,7 @@ import { styled } from "styled-components";
 export const $Main = styled.main`
   margin: 10px 110px; 
   display: block;
+  padding-bottom: 70px;
 `;
 
 export default $Main;