+++ src/components/Main/Reports/Reports.jsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, Suspense } from "react";
 import { useLoaderData, defer, Await } from "react-router-dom";
 import { CSSTransition } from "react-transition-group";
 import { BeatLoader } from "react-spinners";
-import { $Reports, $Fallback, $Children, $Message } from "./Reports.styles.jsx";
+import { $Reports, $Fallback, $Children } from "./Reports.styles.jsx";
 import ReportsForm from "./ReportsForm/ReportsForm.jsx";
 import Countries from "./Countries/Countries.jsx";
 import Documents from "./Documents/Documents.jsx";